cattle_face_recognition/requirement.txt
tensorboard
easydict
mxnet
onnx
sklearn
opencv-python
pytorch